+++ src/main/scala/dxWDL/Main.scala
@@ -202,7 +202,7 @@ object Main extends App {
         val trgName: String = Utils.replaceFileSuffix(wdlSourceFile, ".ir.yaml")
         val trgPath = Utils.appCompileDirPath.resolve(trgName).toFile
         val yo = IR.yaml(irNs)
-        val humanReadable = yo.prettyPrint
+        val humanReadable = IR.prettyPrint(yo)
         val fos = new FileWriter(trgPath)
         val pw = new PrintWriter(fos)
         pw.print(humanReadable)
@@ -410,7 +410,7 @@ object Main extends App {
         val wdlInputs: Option[Path] = options.get("inputs").map(Paths.get(_))
         (wf, irNs.workflow, wdlInputs) match {
             case (Some(dxwfl), Some(irWf), Some(path)) =>
-                InputFile(cOpt.verbose).apply(dxwfl, irWf, path)
+                InputFile(cOpt.verbose).apply(dxwfl, irNs, path)
                 dxwfl.getId
             case _ => ()
         }